Улучшение теста на внимательность 1

+++ src/main/java/itmo/nick/test/attention/TestOneData.java
@@ -17,6 +17,6 @@ import lombok.*;
 
 public class TestOneData {
 	private final Integer stage;
-	private final Integer time;
+	private final Double time;
 	private final Integer errors;
 }
